Created resources folder for preprocessing code
resources/ directory contains the volcanos.csv file of the Smithsonian
volcanos database from which the volcano id and name are parsed from the
image coordinates, as well as the .ttf file of the font being used to
write text onto the images.

The resources folder is a subdirectory of preprocessing/ because its contents are only ever used by preprocessing/ code.