Created docker-compose.yml file to run api and db as conjoined docker contaiers
Note the new code in common/database.py that tries up to 5 times to connect to the mysql database before exiting. This is currently due to the fact that the mysql docker container takes longer to initalize than the api container, and thus the api container must be forced to wait until the mysql container is finished setting up before attempting a connection.